requirements.txt
certifi==2018.1.18
cymem==1.31.2
cytoolz==0.8.2
dill==0.2.7.1
en-core-web-sm==2.0.0
future==0.16.0
msgpack-numpy==0.4.1
msgpack-python==0.5.6
murmurhash==0.28.0
mysqlclient==1.3.12
numpy==1.14.2
pathlib==1.0.1
plac==0.9.6
preshed==1.0.0
python-telegram-bot==10.0.1
regex==2017.4.5
six==1.11.0
spacy==2.0.11
termcolor==1.1.0
thinc==6.10.2
toolz==0.9.0
tqdm==4.21.0
ujson==1.35
wrapt==1.10.11
